[kernel][arm64] Fix operand sizes in inline asm
The code is using operand formats like %0 that mean the 64-bit
register, but operand values that are 32 bits.  These are supposed to
match, but GCC does not check that they do (Clang does).  In assembly
operands where the 32-bit register is valid, the operand format can be
%w0 to emit the 32-bit register name instead of the 64-bit register
name.  But for instructions like mrs, msr, and tlbi, the register
operand must be a 64-bit register.  So the existing code was emitting
assembly like "msr x0, fpcr" but preparing the value using w0.  This
is probably harmless in all cases because using wN clears the high
bits of the register.  But it's sloppy, and calculating in 64-bit
registers doesn't cost any more.

+++ kernel/arch/arm64/fpu.c
@@ -51,7 +51,9 @@ static void arm64_fpu_load_state(struct thread *t)
                      "ldp     q30, q31, [%0, #(15 * 32)]\n"
                      "msr     fpcr, %1\n"
                      "msr     fpsr, %2\n"
-                     :: "r"(fpstate->regs), "r"(fpstate->fpcr), "r"(fpstate->fpsr));
+                     :: "r"(fpstate->regs),
+                        "r"((uint64_t)fpstate->fpcr),
+                        "r"((uint64_t)fpstate->fpsr));
 }
 
 static void arm64_fpu_save_state(struct thread *t)
@@ -87,7 +89,7 @@ static void arm64_fpu_save_state(struct thread *t)
 /* save fpu state if the thread had dirtied it and disable the fpu */
 void arm64_fpu_context_switch(struct thread *oldthread, struct thread *newthread)
 {
-    uint32_t cpacr = ARM64_READ_SYSREG(cpacr_el1);
+    uint64_t cpacr = ARM64_READ_SYSREG(cpacr_el1);
     if (is_fpu_enabled(cpacr)) {
         LTRACEF("saving state on thread %s\n", oldthread->name);
 
@@ -107,7 +109,7 @@ void arm64_fpu_exception(struct arm64_iframe_long *iframe, uint exception_flags)
     /* only valid to be called if exception came from lower level */
     DEBUG_ASSERT(exception_flags & ARM64_EXCEPTION_FLAG_LOWER_EL);
 
-    uint32_t cpacr = ARM64_READ_SYSREG(cpacr_el1);
+    uint64_t cpacr = ARM64_READ_SYSREG(cpacr_el1);
     DEBUG_ASSERT(!is_fpu_enabled(cpacr));
 
     /* enable the fpu */
@@ -119,4 +121,3 @@ void arm64_fpu_exception(struct arm64_iframe_long *iframe, uint exception_flags)
     if (likely(t))
         arm64_fpu_load_state(t);
 }
-
